add sender to JSON generated by GetJson()

+++ WavesCS/Node.cs
@@ -111,7 +111,7 @@ namespace WavesCS
             else
             {
                 var tx = GetObject($"transactions/info/{assetId}");
-                if ((TransactionType)tx.GetInt("type") != TransactionType.Issue)
+                if ((TransactionType)tx.GetByte("type") != TransactionType.Issue)
                     throw new ArgumentException("Wrong asset id (transaction type)");
 
                 asset = new Asset(assetId, tx.GetString("name"), tx.GetByte("decimals"));
@@ -153,7 +153,7 @@ namespace WavesCS
             var typeId = TransactionTypeId(typeof(T));
 
             return GetTransactionsByAddress(address, limit)
-                .Where(tx => (TransactionType)tx.GetInt("type") == typeId)
+                .Where(tx => (TransactionType)tx.GetByte("type") == typeId)
                 .Select(Transaction.FromJson)
                 .Cast<T>()
                 .ToArray();
